Rename rearrangeable list and item.
