+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/DrawableLyricEditList.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Game.Graphics.Containers;
+using osu.Game.Rulesets.Karaoke.Objects;
+
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
+{
+    public class DrawableLyricEditList : OsuRearrangeableListContainer<Lyric>
+    {
+        protected override OsuRearrangeableListItem<Lyric> CreateOsuDrawable(Lyric item)
+            => new DrawableLyricEditListItem(item);
+    }
+}
